+++ app/src/main/java/com/example/skysync/data/remote/WeatherRemoteDataSource.kt
@@ -5,7 +5,7 @@ import com.example.skysync.models.ForecastWeatherResponse
 import kotlinx.coroutines.flow.Flow
 
 interface WeatherRemoteDataSource {
-    suspend fun getCurrentWeather(lat: Long?, lon: Long?, language: String, unit: String): Flow<CurrentWeatherResponse>
-    suspend fun getForecast(lat: Long?, lon: Long?, language: String, unit: String): Flow<ForecastWeatherResponse>
+    suspend fun getCurrentWeather(lat: Double?, lon: Double?, language: String, unit: String): Flow<CurrentWeatherResponse>
+    suspend fun getForecast(lat: Double?, lon: Double?, language: String, unit: String): Flow<ForecastWeatherResponse>
 
 }
